move multiplatform code to the common target
