+++ library/src/commonMain/kotlin/dev/nesk/akkurate/Configuration.kt
