+++ library/src/commonMain/kotlin/dev/nesk/akkurate/Path.kt
