+++ library/src/commonMain/kotlin/dev/nesk/akkurate/ValidationResult.kt
