+++ library/src/commonMain/kotlin/dev/nesk/akkurate/Validator.kt
