+++ library/src/commonMain/kotlin/dev/nesk/akkurate/accessors/Iterable.kt
