+++ library/src/commonMain/kotlin/dev/nesk/akkurate/constraints/Constraint.kt
