+++ library/src/commonMain/kotlin/dev/nesk/akkurate/constraints/ConstraintDescriptor.kt
