+++ library/src/commonMain/kotlin/dev/nesk/akkurate/constraints/ConstraintRegistry.kt
