+++ library/src/commonMain/kotlin/dev/nesk/akkurate/constraints/ConstraintViolationSet.kt
