+++ library/src/commonMain/kotlin/dev/nesk/akkurate/constraints/builders/Map.kt
@@ -21,6 +21,7 @@ import dev.nesk.akkurate.constraints.Constraint
 import dev.nesk.akkurate.constraints.constrainIfNotNull
 import dev.nesk.akkurate.constraints.otherwise
 import dev.nesk.akkurate.validatables.Validatable
+import kotlin.jvm.JvmName
 
 /*
  * !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
