+++ library/src/commonMain/kotlin/dev/nesk/akkurate/validatables/ValidatableCompound.kt
