+++ library/src/commonTest/kotlin/dev/nesk/akkurate/ConfigurationTest.kt
