+++ library/src/commonTest/kotlin/dev/nesk/akkurate/_test/Validatable.kt
