+++ library/src/commonTest/kotlin/dev/nesk/akkurate/accessors/IterableTest.kt
