+++ library/src/commonTest/kotlin/dev/nesk/akkurate/constraints/builders/IterableTest.kt
